backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic-settings>=2.7
python-jose[cryptography]>=3.3
authlib>=1.4
httpx>=0.28
google-api-python-client>=2.160
google-auth>=2.37
openpyxl>=3.1
python-multipart>=0.0.18
itsdangerous>=2.2
